Add channel input and package to release
The previous version did not include the package version when building a release
This change pulls in the template in order to set the package version for the
release that is created. The channel name is needed in order to get the proper
deployment template, so the additional input was added as optional. The channel
defaults to 'Default' as created by Octopus Deploy